bot/requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
web3>=6.15.0
python-telegram-bot>=20.7
py-clob-client>=0.18.0
